requirements.txt
yfinance==0.2.43
pandas==2.2.2 
numpy==2.1.0 
alpaca-py==0.30.1
python-binance==1.0.19